+++ www/client/src/modules/Midi2MusicXML/utils/midiNoteToPitch.ts
@@ -1,0 +1,8 @@
+export function midiNoteToPitch(midiNote: number) {
+  const stepNames = ['C', 'C', 'D', 'D', 'E', 'F', 'F', 'G', 'G', 'A', 'A', 'B'];
+  const alterMap = [0, 1, 0, 1, 0, 0, 1, 0, 1, 0, 1, 0];
+  const step = stepNames[midiNote % 12];
+  const alter = alterMap[midiNote % 12];
+  const octave = Math.floor(midiNote / 12) - 1;
+  return { step, alter, octave };
+}
